% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodetfakepkg3.R
\name{tinycodetfakepkg3}
\alias{tinycodetfakepkg3}
\alias{fun_overwritten}
\alias{fun31}
\alias{fun32}
\alias{\%opover\%}
\alias{\%op31\%}
\alias{\%op32\%}
\alias{rpbase_overwritten}
\alias{rpbase31}
\alias{rpbase32}
\alias{rpbase_overwritten<-}
\alias{rpbase31<-}
\alias{rpbase32<-}
\alias{const_overwritten}
\alias{const31}
\alias{const32}
\title{tinycodetfakepkg3ellaneous functions to help your coding etiquette}
\usage{
fun_overwritten()

fun31()

fun32()

X \%opover\% A

X \%op31\% A

X \%op32\% A

rpbase_overwritten(x)

rpbase31(x)

rpbase32(x)

rpbase_overwritten(x) <- value

rpbase31(x) <- value

rpbase32(x) <- value

const_overwritten

const31

const32
}
\arguments{
\item{X}{whatever}

\item{A}{whatever}

\item{x}{description}

\item{value}{description}
}
\description{
Bla
}
