% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fun11}
\alias{acf}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stats}{\code{\link[stats]{acf}}}

  \item{tinycodetfakepkg1}{\code{\link[tinycodetfakepkg1]{fun11}}}
}}

