% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodetfakepkg3.R
\name{tinycodetfakepkg3}
\alias{tinycodetfakepkg3}
\alias{fun_overwritten}
\alias{fun31}
\alias{fun32}
\alias{\%opover\%}
\alias{\%op31\%}
\alias{\%op32\%}
\title{tinycodetfakepkg3ellaneous functions to help your coding etiquette}
\usage{
fun_overwritten()

fun31()

fun32()

X \%opover\% A

X \%op31\% A

X \%op32\% A
}
\arguments{
\item{X}{whatever}

\item{A}{whatever}
}
\description{
Bla
}
